--- app/src/main/java/com/p2p/wallet/history/interactor/TokenInteractor.kt
@@ -1,15 +1,0 @@
-package com.p2p.wallet.history.interactor
-
-import com.p2p.wallet.history.model.PriceHistory
-import com.p2p.wallet.history.repository.TokenRepository
-
-class TokenInteractor(
-    private val tokenRepository: TokenRepository
-) {
-
-    suspend fun getDailyPriceHistory(sourceToken: String, destination: String, days: Int): List<PriceHistory> =
-        tokenRepository.getDailyPriceHistory(sourceToken, destination, days)
-
-    suspend fun getHourlyPriceHistory(sourceToken: String, destination: String, hours: Int): List<PriceHistory> =
-        tokenRepository.getHourlyPriceHistory(sourceToken, destination, hours)
-}
