+++ app/src/main/java/com/p2p/wallet/main/interactor/MainInteractor.kt
@@ -1,35 +1,24 @@
 package com.p2p.wallet.main.interactor
 
 import com.p2p.wallet.R
-import com.p2p.wallet.common.crypto.Base64Utils
-import com.p2p.wallet.history.model.TransactionConverter
-import com.p2p.wallet.history.model.TransactionType
 import com.p2p.wallet.infrastructure.network.provider.TokenKeyProvider
 import com.p2p.wallet.main.model.Token
 import com.p2p.wallet.main.model.TransactionResult
 import com.p2p.wallet.rpc.repository.FeeRelayerRepository
 import com.p2p.wallet.rpc.repository.RpcRepository
-import com.p2p.wallet.user.repository.UserLocalRepository
 import com.p2p.wallet.utils.toPublicKey
 import org.p2p.solanaj.kits.TokenTransaction
-import org.p2p.solanaj.kits.transaction.CloseAccountDetails
-import org.p2p.solanaj.kits.transaction.SwapDetails
-import org.p2p.solanaj.kits.transaction.TransactionTypeParser
-import org.p2p.solanaj.kits.transaction.TransferDetails
-import org.p2p.solanaj.kits.transaction.UnknownDetails
 import org.p2p.solanaj.model.core.Account
 import org.p2p.solanaj.model.core.PublicKey
 import org.p2p.solanaj.model.core.Transaction
 import org.p2p.solanaj.model.core.TransactionInstruction
 import org.p2p.solanaj.programs.SystemProgram
 import org.p2p.solanaj.programs.TokenProgram
-import timber.log.Timber
 import java.math.BigInteger
 
 class MainInteractor(
     private val rpcRepository: RpcRepository,
     private val feeRelayerRepository: FeeRelayerRepository,
-    private val userLocalRepository: UserLocalRepository,
     private val tokenKeyProvider: TokenKeyProvider
 ) {
 
@@ -71,10 +60,6 @@ class MainInteractor(
 
         val signers = listOf(Account(tokenKeyProvider.secretKey))
         transaction.sign(signers)
-
-        val serializedMessage = transaction.serialize()
-        val base64Trx: String = Base64Utils.encode(serializedMessage)
-        Timber.d("### base64 $base64Trx")
 
         val signature = transaction.getSignature().orEmpty()
 
@@ -153,15 +138,16 @@ class MainInteractor(
         val signers = listOf(Account(tokenKeyProvider.secretKey))
         transaction.sign(signers)
 
-        val serializedMessage = transaction.serialize()
-        val base64Trx: String = Base64Utils.encode(serializedMessage)
-        Timber.d("### serialized message: $base64Trx")
-
         val signature = transaction.getSignature().orEmpty()
 
+        val recipientPubkey = if (address.equals(destinationAddress)) {
+            address.toBase58()
+        } else {
+            destinationAddress.toBase58()
+        }
         val transactionId = feeRelayerRepository.sendSplToken(
             senderTokenAccountPubkey = token.publicKey,
-            recipientPubkey = address.toBase58(),
+            recipientPubkey = recipientPubkey,
             tokenMintPubkey = token.mintAddress,
             authorityPubkey = tokenKeyProvider.publicKey,
             lamports = lamports,
@@ -196,74 +182,4 @@ class MainInteractor(
 
         throw IllegalStateException("Wallet address is not valid")
     }
-
-    suspend fun getHistory(publicKey: String, before: String?, limit: Int): List<TransactionType> {
-        val signatures = rpcRepository.getConfirmedSignaturesForAddress(
-            publicKey.toPublicKey(), before, limit
-        ).map { it.signature }
-
-        return rpcRepository.getConfirmedTransactions(signatures)
-            .mapNotNull { response ->
-                val signature = response.transaction.signatures.firstOrNull()
-                val data = TransactionTypeParser.parse(signature, response)
-
-                val swap = data.firstOrNull { it is SwapDetails }
-                if (swap != null) {
-                    return@mapNotNull parseSwapDetails(swap as SwapDetails)
-                }
-
-                val transfer = data.firstOrNull { it is TransferDetails }
-                if (transfer != null) {
-                    return@mapNotNull parseTransferDetails(
-                        transfer as TransferDetails,
-                        publicKey,
-                        tokenKeyProvider.publicKey
-                    )
-                }
-
-                val close = data.firstOrNull { it is CloseAccountDetails }
-                if (close != null) {
-                    return@mapNotNull parseCloseDetails(close as CloseAccountDetails)
-                }
-
-                val unknown = data.firstOrNull { it is UnknownDetails }
-                TransactionConverter.fromNetwork(unknown as UnknownDetails)
-            }
-            .sortedByDescending { it.date.toInstant().toEpochMilli() }
-    }
-
-    private fun parseSwapDetails(details: SwapDetails): TransactionType? {
-        val sourceData = userLocalRepository.getTokenData(details.mintA) ?: return null
-        val destinationData = userLocalRepository.getTokenData(details.mintB) ?: return null
-
-        if (sourceData.mintAddress == destinationData.mintAddress) return null
-
-        val destinationRate = userLocalRepository.getPriceByToken(destinationData.symbol)
-        val source = tokenKeyProvider.publicKey
-        return TransactionConverter.fromNetwork(details, sourceData, destinationData, destinationRate, source)
-    }
-
-    private fun parseTransferDetails(
-        transfer: TransferDetails,
-        directPublicKey: String,
-        publicKey: String
-    ): TransactionType {
-        val symbol = if (transfer.isSimpleTransfer) Token.SOL_SYMBOL else findSymbol(transfer.mint)
-        val rate = userLocalRepository.getPriceByToken(symbol)
-
-        val mint = if (transfer.isSimpleTransfer) Token.SOL_MINT else transfer.mint
-        val source = userLocalRepository.getTokenData(mint)!!
-
-        return TransactionConverter.fromNetwork(transfer, source, directPublicKey, publicKey, rate)
-    }
-
-    private suspend fun parseCloseDetails(details: CloseAccountDetails): TransactionType {
-        val accountInfo = rpcRepository.getAccountInfo(details.account.toPublicKey())
-        val info = TokenTransaction.parseAccountInfoData(accountInfo, TokenProgram.PROGRAM_ID)
-        val symbol = findSymbol(info?.mint?.toBase58().orEmpty())
-        return TransactionConverter.fromNetwork(details, symbol)
-    }
-
-    private fun findSymbol(mint: String): String =
-        if (mint.isNotEmpty()) userLocalRepository.getTokenData(mint)?.symbol.orEmpty() else ""
 }
